+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+* Rewrite to the same layout as the `alloc` crate.
+
 # 0.2.1 - 2019-01-28
 
 * Fix a bug that can be compiled with the stable compiler when both `std` and `alloc` are specified.
@@ -8,7 +10,7 @@
 
 # 0.2.0 - 2019-01-27
 
-* Change APIs. You can also use it like [this example](examples/std-shim).
+* Change APIs. You can also use it like [this example](https://github.com/taiki-e/alloc-shim/tree/v0.2.0/examples/std-shim).
 
 * Change lib name to `alloc`.